feat: move connection inbox into modal triggered by header CTA chip
Replace the inline pending-suggestions banner with a pulsing "Review X
new suggestions" chip in the page header (next to the rank-movies pill).
Clicking the chip opens a dedicated modal containing the same 4-button
response grid, making the call-to-action discoverable and the action
itself focused. The pending connection requests banner stays inline.
